refactor(phaseC): split EepromManager into persistence/eeprom/ domain files
Splits the 651-line EepromManager.{h,cpp} monolith (Phase A audit finding 4,
the largest single-responsibility violation in the persistence subsystem)
into 8 domain files under persistence/eeprom/: Core (NVS primitives + shared
state), Identity, Role, Security, Peers, Diagnostics, Enrollment, and
DeviceConfig. Every function keeps its exact name, signature, and
lattice::eeprom:: namespace -- this is a pure move, not a rename.

Each consumer now includes only the domain header(s) it actually calls into
instead of the old catch-all EepromManager.h. Found and fixed two gaps the
task brief's consumer table missed: PeerRegistry.h (not .cpp) held the old
include for EEPROM_SIZES::MAX_PEERS, and ReplayCache.h (absent from the
brief entirely) calls eeprom::saveBootEpoch() directly.

53 TEST_F cases moved verbatim from test_eeprom_manager.cpp into 8 new
per-domain test files (name-for-name diffed against the original for an
exact match). Unit (289) + e2e (41) = 330 tests, matching the pre-task count.
