+++ firmware/main/main.cpp
@@ -7,7 +7,10 @@
 #include "src/hardware/input/Button.h"
 #include "src/error/Error.h"
 #include "src/error/ErrorCore.h"
-#include "src/persistence/EepromManager.h"
+#include "src/persistence/eeprom/EepromCore.h"
+#include "src/persistence/eeprom/EepromIdentity.h"
+#include "src/persistence/eeprom/EepromRole.h"
+#include "src/persistence/eeprom/EepromPeers.h"
 #include "src/app/BootManager.h"
 #include "src/app/DisplayManager.h"
 #include "src/app/ButtonHandler.h"
