+++ firmware/main/src/mesh/Enrollment.cpp
@@ -1,6 +1,8 @@
 #include "Enrollment.h"
 #include "MeshCrypto.h"
-#include "src/persistence/EepromManager.h"
+#include "src/persistence/eeprom/EepromIdentity.h"
+#include "src/persistence/eeprom/EepromSecurity.h"
+#include "src/persistence/eeprom/EepromEnrollment.h"
 #include "src/logging/Logger.h"
 #include "src/error/Error.h"
 #include "../../lib/lattice-protocol/c/mesh_message.h"
