+++ firmware/main/src/mesh/PeerRegistry.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 #include <cstring>
-#include "src/persistence/EepromManager.h"
+#include "src/persistence/eeprom/EepromCore.h"
 #include "src/network/MacAddress.h"
 #include "../../project_config.h"
