+++ firmware/main/src/persistence/eeprom/EepromEnrollment.h
@@ -1,0 +1,9 @@
+#ifndef LATTICE_EEPROM_ENROLLMENT_H
+#define LATTICE_EEPROM_ENROLLMENT_H
+namespace lattice {
+namespace eeprom {
+bool loadEnrolledFlag();
+void saveEnrolledFlag(bool enrolled);
+} // namespace eeprom
+} // namespace lattice
+#endif
