+++ tests/e2e/harness/SimNode.cpp
@@ -10,7 +10,9 @@
 #include "src/error/ErrorCore.h"
 #include "src/error/Error.h"
 #include "src/hardware/output/Led.h"
-#include "src/persistence/EepromManager.h"
+#include "src/persistence/eeprom/EepromCore.h"
+#include "src/persistence/eeprom/EepromRole.h"
+#include "src/persistence/eeprom/EepromIdentity.h"
 #include "src/logging/Logger.h"
 #include "project_config.h"
 #include <cstring>
