+++ tests/mocks/nvs_flash.h
@@ -3,7 +3,7 @@
 // header. Only nvs_flash_init()/nvs_flash_erase() live here (main.cpp's boot
 // sequence, Phase I Task 4); the handle-based get/set API is in nvs.h.
 // Not exercised by host unit/e2e tests today (main.cpp isn't part of the host
-// build), but kept in step with the real header split so EepromManager.h's
+// build), but kept in step with the real header split so EepromCore.h's
 // #include set matches the ESP-IDF target 1:1.
 #include "esp_err.h"
 #include "nvs.h"
